nfs-utils: Tidy up libnfsidmap exported symbols list and API
Prior to its merger with nfs-utils the libnfsidmap library leaked
a lot of internal functions and variables that were not defined
in its API, some of which were used by its various plugins.
This created an accidentally large ABI which restricts attempts to
clean up the code base.

These patches aim to trim the ABI down, whilst adding to the API
only the minimal required to still support both bundled and 3rd
party plugins.

Signed-off-by: Justin Mitchell <jumitche@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>