Add headers to support libjunction
libjunction contains support for reading or updating a junction
that resides on an NFS server. It is used by mountd to read junction
information from an exported filesystem, and by the nfsref command
line tool. These will be introduced in subsequent patches.

The source code for libjunction used to be maintained in
fedfs-utils, but fedfs-utils is being retired. Although interest in
FedFS junctions has waned, NFS basic junctions are still appealing
enough that we've decided we want to continue supporting them.
Therefore this patch moves the non-LDAP parts of libjunction into
nfs-utils, to enable nfs-utils to also support the user-visible
pieces of NFS basic junctions.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>