Remove fedfs_admin.h
nfs-utils is not going to provide an implementation of the FedFS
ADMIN protocol, so there's no need to include the entire .x file
for that protocol.

The only piece needed (for the moment) is the enum that defines
the FEDFS_ERR codes, which are used by the internal and external
libjunction APIs.

Suggested-by: Neil Brown <neilb@suse.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>