mount.nfs: minorversion setting is being ignored with the -t flag
mount -t nfs or mount -t nfs4 and setting minorversion
should set the v4 minor version. This patch adds a few
checks to make sure the minor version is set.

The patch also translate the minorversion= option
to vers=4.x in the arguments passed to the kernel
when the kernel is later then v3.4.

Signed-off-by: Steve Dickson <steved@redhat.com>