gssd: Add "srchost=" upcall parameter
The callback client used by NFSv4.0 servers to send CB calls has to
use a source principal that is the same as the target principal
that the client used to establish the forward channel.

In multi-homed server set-ups, the domain part of the principal's
hostname may not be the same as server's DNS domain. So gssd can
no longer assume that. The kernel can scrape that domain off the
forward channel's principal and pass that up to gssd.

This patch adds a new parameter, "srchost", to the kernel upcall.
When the kernel presents this new parameter in an upcall, gssd will
use it, along with the "service" parameter, to construct the service
principal for the keytab lookup.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>