nfsd: track when the user explicitly requested version 4.0
Since both 'rpc.nfsd -V4' and 'rpc.nfsd -V4.0' alter bit 0 in 'minorversset'
and 'minorvers', it's not possible to know whether the user explicitly
requested version 4.0 or not.  Add another variable 'force4dot0' to track
when the user explicitly requested version 4.0 so that we can write '4.0'
to /proc/fs/nfsd/versions on kernel versions 4.11 and higher.  On earlier
kernels, we'll still write '4' as in the past.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>