systemd/rpc-statd.service - add IgnoreOnIsolate=yes
If you run systemctl isolate multi-user.target
then all services not needed by multi-user.target are stopped.
This currently includes rpc-statd, which is a problem.

Systemd doesn't know that rpc-statd.service is needed by
multi-user.target, or more accurately nfs-client.target, because
sometimes it isn't.

It is started dynamically by mount.nfs when a v3 or v2 filesystem
is mounted (not when v4 is mounted) - so systemd doesn't see the connection.

mount units default to IgnoreOnIsolate=yes, as mounts are expected to be
left mounted when you run "systemctl isolate ....".  As rpc-statd.service
is started by a mount, and needed for some mounts, it makes sense for it
to declare IgnoreOnIsolate=yes too.

With this declaration in place rpc.statd remains running when
systemctl isolate multi-user.target is run.

Signed-off-by: NeilBrown <neilb@suse.com>
Signed-off-by: Steve Dickson <steved@redhat.com>