nfsdcld: add a facility for migrating from older client tracking methods
This patch adds a one-time "upgrade" when using nfsdcld for the first
time.  During initialization of the database, we try to copy records
from the old nfdscltrack database and from the legacy v4recovery
directory (although we don't expect to find records in both places).

nfsdcltrack stores the client name string so the handling of those
records is straightforward.  Legacy records are md5 hashes of the client
name string, so we prefix them with "hash:" as way to let knfsd know
that this is a legacy record.  knfsd will need a fallback check in the
event that it cannot find a reclaim record using the client name string.

Upon receipt of the first "GraceDone" upcall we clean out any records
from the old nfsdcltrack database and delete any subdirectories under
the v4recovery directory.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>