mountstats: Fix nfsstat command to handle RPC iostats version >= 1.1
Later kernels with RPC iostats version >= 1.1 have an additional errors
count for each op.  Lengthen the array of values created inside
DeviceData and then in __parse_rpc_line just zero this value out for
prior kernels where this count is not present.  The count is not used
for nfsstat, but this keeps DeviceData consistent with the new count
as well as proper functioning of accumulate_iostats.

Before this patch, nfsstat will backtrace on a kernel with RPC iostats
version >= 1.1 due to the fixed array inside DeviceData.  This patch
fixes this backtrace and also allows nfsstat to work with these new
kernels.

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>