locktest: Makefile.am: remove host compiler costraint
Currently locktest can be built only for host because CC_FOR_BUILD is
specified as CC, but this leads to build failure when passing CFLAGS not
available on host gcc(i.e. -mlongcalls) and most of all locktest would
be available on target systems the same way as rpcgen etc. So remove CC
and LIBTOOL assignments.

Reviewed-by: Petr Vorel <petr.vorel@gmail.com>
Tested-by: Petr Vorel <petr.vorel@gmail.com>
Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>
Signed-off-by: Steve Dickson <steved@redhat.com>