Allow compilation to succeed with -fno-common
When compiled with -fno-common, global variables that are declared
multple times cause an error.  With -fcommon (the default), they are
merged.

Declaring such variable multiple times is probably not a good idea, and
is definitely not necessary.

This patch changes all the global variables defined in include files to
be explicitly "extern", and where necessary, adds the variable
declaration to a suitable .c file.

To test, run
  CFLAGS=-fno-common ./configure
  make

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>