mount: report error if multiple version specifiers are given.
The NFS version can be requested with multiple different options:
  v2 v3 v4 v4.x vers=x nfsvers=

If multiple versions are given with different options, the choice of
which wins is quite ideosyncratic.  It certainly isn't simple "last one
wins" as with some other options.

Rather than providing a coherent rule, simply make multiple version
specifiers illegal.

This requires enhancing po_contains_prefix() to be able to look beyond
the first match, it see if there are multiple matches with the same
prefix, as well as checking all prefixes to see if more than one
matches.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>