mount: fix parsing of default options
A recent patch to change configfile.c to use parse_opt.c contained code
which was intended to remove all "default*" options from the list before
that could be passed to the kernel.  This code didn't work, so default*
options WERE passed to the kernel, and the kernel complained and failed
the mount attempt.

A more recent patch attempted to fix this by not including the
"default*" options in the option list at all.  This resulting in
global-default defaults over-riding per-mount or per-server defaults.

This patch reverse the "more recent" patch, and fixes the original patch
by providing correct code to remove all "default*" options before the
kernel can see them.

Fixes: 88c22f924f1b ("mount: convert configfile.c to use parse_opt.c")
Fixes: 8142542bda28 ("mount: parse default values correctly")
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>