mountd: reject unknown client IP when !use_ipaddr.
When use_ipaddr is not in effect, an auth_unix_ip lookup
request from the kernel for an unknown client will be
rejected.  When it IS in effect, these requests are
always granted with the IP address being mapped to a
string form of the address, preceded by a '$'.

This is inconsistent behaviour and could present a small
information leak. It means that, for example, a SETCLIENT
NFSv4 request may or may not succeed depending on an
internal setting in rpc.mountd.

This is easily rectified by always checking if the
client is known.

Signed-off-by: NeilBrown <neil@brown.name>
Signed-off-by: Steve Dickson <steved@redhat.com>