mountd: add logging of NFSv4 clients attaching and detaching.
NFSv4 does not have a MOUNT request like NFSv3 does (via the MOUNT
protocol).  So these cannot be logged.
NFSv4 does have SETCLIENTID and EXCHANGE_ID.  These are indirectly
visible though changes in /proc/fs/nfsd/clients.
When a new client attaches, a directory appears.  When the client
detaches, through a timeout (v4.0) or DESTROY_SESSION (v4.1+)
the directory disappears.

This patch adds tracking of these changes using inotify, with log
messages when a client attaches or detaches.

Unfortuantely clients are created in two steps, the second being a
confirmation.  This results in a temporary client appearing and
disappearing.  It is not possible (in Linux 5.10) to detect the
unconfirmed client, so extra attach/detach messages are generated.

This patch also moves some cache* function declarations into a header
file, and makes a few related changes to #includes.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>