Move declaration of etab and rmtab into libraries
There are two global "struct stat_paths" structures: etab and rmtab.
They are currently needed by some library code so any program which is
linked with that library code needs to declare the structures even if it
doesn't use the functionality.  This is clumsy and error-prone.

Instead: have the library declare the structure and put the definition
in a header file.  Now programs only need to know about these structures
if they use the functionality.

'rmtab' is now declared in libnfs.a (rmtab.c).  'etab' is declared in
export.a (xtab.c).

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>