systemd/Makefile: Drop exlicit setting of unit_dir
The used variable is named unitdir in configure.ac and
systemd/Makefile.am otherwise but is used in a single place as unit_dir.

The setting has no effect, but if later commits would use a base to the
systemd unit files directory for installing further files this would
void the possibility to explicitly set a systemd unit files directory
via configure with --with-systemd=unit-dir-path.

Fixes: 0fbf91a4fd90 ("Include systemd unit files in "dist" and "install".")
Signed-off-by: Salvatore Bonaccorso <carnil@debian.org>
Signed-off-by: Steve Dickson <steved@redhat.com>