Manpages: Fix man page syntax errors
In idmapd.conf.5, there is a line of what should be literal text
beginning with ".", which makes it an (invalid) command.  It can be
escaped, but then there will be a space before it.  Instead, Move it
to the previous line and use the .BR macro so there's no space.

In idmapd.man, the .I (italic) macro is used.  However, this manual
page uses the mdoc macro package that does not include it.  Use the
.Em (emphasis) macro instead.

In nfsmount.conf.man, the first line should be a comment but it is
actually an invalid command.  Fix it to be a comment.

Signed-off-by: Ben Hutchings <benh@debian.org>
Signed-off-by: Salvatore Bonaccorso <carnil@debian.org>
Signed-off-by: Steve Dickson <steved@redhat.com>